formalize
